# api-gateway/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  booking.CreateServiceRequest:
    properties:
      description:
        type: string
      duration:
        description: Duration in minutes
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
  booking.ListServicesResponse:
    properties:
      services:
        items:
          $ref: '#/definitions/booking.ServiceResponse'
        type: array
    type: object
...
  booking.ServiceResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  booking.UpdateServiceRequest:
    properties:
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  models.UUser:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  users.GetProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        description: Should be 'customer', 'provider', or 'admin'
        type: string
    type: object
...
  users.UserResponseU:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway of On-Demand Car Wash Service
  title: On-Demand Car Wash Service
  version: "1.0"
paths:
...
  /admin/services/{id}:
    get:
      description: Retrieves service details by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ServiceResponse'
        "400":
          description: Invalid service ID
          schema:
            type: string
        "404":
          description: Service not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a service by ID
      tags:
      - services
...
  /customer/profile:
    get:
      description: Retrieves user profile by ID
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.GetProfileResponse'
        "400":
          description: Invalid user id
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets user
      tags:
      - user
    put:
      description: Updates user info by ID
      parameters:
      - description: New user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UUser'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponseU'
        "400":
          description: Invalid user id or data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates user
      tags:
      - user
...
  /customer/services:
    get:
      description: Retrieves a list of services with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page limit
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListServicesResponse'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Lists all services with pagination
      tags:
      - services
  /customer/services/{id}:
    get:
      description: Retrieves service details by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ServiceResponse'
        "400":
          description: Invalid service ID
          schema:
            type: string
        "404":
          description: Service not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a service by ID
      tags:
      - services
  /customer/services/popular:
    get:
      description: Retrieves service details
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListServicesResponse'
        "400":
          description: Invalid service ID
          schema:
            type: string
        "404":
          description: Service not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a popular service
      tags:
      - services
...
  /provider/services:
    get:
      description: Retrieves a list of services with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page limit
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListServicesResponse'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Lists all services with pagination
      tags:
      - services
    post:
      description: Creates a new service and returns the service details
      parameters:
      - description: Service details
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/booking.CreateServiceRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ServiceResponse'
        "400":
          description: Invalid service data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a new service
      tags:
      - services
  /provider/services/{id}:
    put:
      description: Updates service details by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated service details
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/booking.UpdateServiceRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ServiceResponse'
        "400":
          description: Invalid service data or ID
          schema:
            type: string
        "404":
          description: Service not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates a service by ID
      tags:
      - services
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"